pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.1",
    "google-genai>=1.62.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "numpy>=2.4.2",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]
